Bypass the login to Gerrit as it prevents Digest auth from working properly
Had to cancel the 2 calls (attempt & fallback) to /login/ as it seems to
prevent the HTTP Digest auth from working properly. As we're behind a reverse
proxy, the password & Gerrit HTTP password may be different.
Gerrit API requires the Gerrit HTTP password, while the access to /login/
requires the password from the reverse proxy.

Somehow, when attempting to login on /login/, failing, then making the request
to /a/, the Digest authentication wasn't kicking in. Didn't bother to
understand why at the moment, but not calling /login/ let's the HTTP Digest
auth proceed normally.